Initial checkin from Jerky post CAX 2019... here are his release notes.
Level 13 - 
The flow of Ions out of the cannon is a steady stream now which times out/just looks better. I thought making it offset timing wise would work better to make the player think more about weaving through Ion shots...but it just didn't feel right after playing it for weeks now. A steady stream seems like a better way to give them something to learn with. It also makes getting back up to the top a little trickier as an added bonus ;)


Level 16 - 
I changed the two Pyroids(95 and 95) before the Reactor to a movement pattern that is a slightly different version of the one on Level 18. It just flows better this way with the IonCannon shots as well.


Level 22 -
I added another obstacle IonCannon(1) just before Cannon19 (the one that fires shots) in the hallway gag that leads to the Blue Key. It felt like something was missing up there after getting past the first two (Cannon21 then 20). It also rewards players who get through faster with an easier Pyroid setup in the lower hallway on the way back to the Cyan Key (the Pyroids line up just like they did the first time through with plenty of breathing room). If you mess up and have to dodge an extra wave of Ions in this Ion Cannon hallway you end up in between the pairs of Ions in the lower hallway which is a bit harder to get through timing wise. Screw up too many times in the Ion Cannon hallway and you will have to wait longer for a window to get through the Pyroid gag at the bottom of the maze. I also moved the Flashing Transporter leading to L24 down to the bottom of the last vertical shaft on the right below IonCannon23.


Level 23 - 
I changed the IonCannon23 program again to one that flows better in the hallway at the bottom of the vertical shaft (syncs up with vertical shots from above better spacing wise). It also now blocks the player from moving forward right before Cannon23 on the way to the flashing transporter. Now they have to go up and over the shots as they come out of the transporter before they have to get underneath the cannon as it lowers again (that's why the up arrow is there to hint them in the right direction).